fix/#10 : inner class로 수정
