+++ app/src/main/java/com/sopt/now/compose/domain/entity/response/ResponseUserList.kt
@@ -1,6 +1,5 @@
 package com.sopt.now.compose.domain.entity.response
 
-import com.sopt.now.compose.data.remote.dto.response.reqres.Data
 import kotlinx.collections.immutable.ImmutableList
 
 data class ResponseUserList(
